refactor: move data layer interfaces to folders
issue: #8"
