+++ src/data/protocols/db/load-account-by-email-repository.ts
@@ -1,4 +1,4 @@
-import { AccountModel } from '../usecases/add-account/db-add-account-protocols'
+import { AccountModel } from '../../usecases/add-account/db-add-account-protocols'
 
 export interface LoadAccountByEmailRepository {
   load: (email: string) => Promise<AccountModel>
