+++ src/data/protocols/db/log-error-repository.ts
