+++ src/infra/cryptography/bcrypt-adapter.ts
@@ -1,5 +1,5 @@
 import bcrypt from 'bcrypt'
-import { Encrypter } from '../../data/protocols/encrypter'
+import { Encrypter } from '../../data/protocols/criptography/encrypter'
 
 export class BcryptAdapter implements Encrypter {
   constructor (private readonly salt: number) { }
